% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_script.R
\docType{data}
\name{master_script}
\alias{master_script}
\alias{redcap_uri}
\title{The parent script of the DSMC project}
\format{
An object of class \code{character} of length 1.
}
\usage{
redcap_uri
}
\description{
This script is responsible for running the DSMC report.
This project is being adapted to a package, which will be available at \url{https://github.com/jclutton/dsmreport}
Written by Jon Clutton
Copyright at University of Kansas 2023
}
\section{Instructions}{

This code will require some editing on the user's part.
\enumerate{
\item Be sure to review the instructions at \url{https://github.com/jclutton/DSMC}
\item Edit the information in this script. Any info that needs to be edited has EDIT in the comments.
}
}

\keyword{datasets}
