% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{open_consort}
\alias{open_consort}
\title{Create Open Consort Image}
\usage{
open_consort(consort_data = NULL, meta_data = NULL, filepath = NULL)
}
\arguments{
\item{consort_data}{A list of dataframes from \code{\link[=prep_consort]{prep_consort()}}
\itemize{
\item \code{screened}: All screened participants that have a \code{scrn_date}
\item \code{not_screened}: Participants in the database without a \code{scrn_date}, presumably to be screened
\item \code{screen_failed}: Participants who screen-fail before consenting, \code{screening_status} = "Screen-Fail"
\item \code{eligible}: Participants who pass pre-consent screening, \code{screening_status} = "Eligible"
\item \code{on_hold}: Participants who are on-hold before consent, \code{screening_status} = On-hold
\item \code{still_screening}: Participants who are presumable still screening because \code{scrn_date} is not blank but \code{screening_status} is blank
\item \code{screen_fail_string}: String of pre-consent screenfail reasons and numbers for the consort diagram
\item \code{consented}: Participants with a \code{consent_date}
\item \code{expected_to_consent}: Participants whose \code{screening_status} = "Eligible" but \code{consent_date} is blank
\item \code{randomized}: Participants who have a \code{rand_date}
\item \code{screen_failed_after_consent}: Participants whose \code{baseline_status} = Screen Fail
\item \code{in_baseline}: Participants who have a \code{consent_date} but do not have a \code{rand_date}
\item \code{screen_fail_baseline_string}: String of baseline screen-fail reasons and numbers for the consort diagram
\item \code{active}: Participants whose \code{intervention_status} is Actively on intervention or is blank
\item \code{completed}: Participants whose \code{intervention_status} is Successfully completed
\item \code{withdrawn_testing}: Participants whose \code{intervention_status} is Withdrawn willing to complete testing
\item \code{withdrawn_no_testing}: Participants whose \code{intervention_status} is Withdrawn unwilling to complete testing
\item \code{lost}: Participants whose \code{intervention_status} is Lost to Follow-up
@inheritParams prep_consort
@param filepath The path to a folder to save a svg image of the consort diagram
}

@returns Images as .png files
\itemize{
\item Post-consent diagram
}

@importFrom REDCapR checkbox_choices
@import dplyr
@import ggplot2

@section Notes:
20240418: Idea for later. Have constants with the subtract or additions that change depending on the size of the grid so that how much I'm subtracting
is relative to image. Also, group 1 and 2 start central. Then group 3 and four move to the sides. JC}
}
\description{
This function uses the dataframes created in \code{\link[=prep_consort]{prep_consort()}} to create an
open consort diagram. The consort diagram is not a standard consort diagram, as it is
designed for ongoing trials; however, it conveys the same information.
}
