% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{prep_consort}
\alias{prep_consort}
\title{Prep Consort}
\usage{
prep_consort(df = NULL, meta_data = NULL)
}
\arguments{
\item{df}{A dataframe from \code{\link[=get_data]{get_data()}} with all relevant variables}

\item{meta_data}{A dataframe with the project REDCap dictionary from \code{\link[=get_data]{get_data()}}}
}
\value{
Dataframes returned in a list
\itemize{
\item \code{screened}: All screened participants that have a \code{scrn_date}
\item \code{not_screened}: Participants in the database without a \code{scrn_date}, presumably to be screened
\item \code{screen_failed}: Participants who screen-fail before consenting, \code{screening_status} = "Screen-Fail"
\item \code{eligible}: Participants who pass pre-consent screening, \code{screening_status} = "Eligible"
\item \code{on_hold}: Participants who are on-hold before consent, \code{screening_status} = On-hold
\item \code{still_screening}: Participants who are presumable still screening because \code{scrn_date} is not blank but \code{screening_status} is blank
\item \code{screen_fail_string}: String of pre-consent screenfail reasons and numbers for the consort diagram
\item \code{consented}: Participants with a \code{consent_date}
\item \code{expected_to_consent}: Participants whose \code{screening_status} = "Eligible" but \code{consent_date} is blank
\item \code{randomized}: Participants who have a \code{rand_date}
\item \code{screen_failed_after_consent}: Participants whose \code{baseline_status} = Screen Fail
\item \code{in_baseline}: Participants who have a \code{consent_date} but do not have a \code{rand_date}
\item \code{screen_fail_baseline_string}: String of baseline screen-fail reasons and numbers for the consort diagram
\item \code{active}: Participants whose \code{intervention_status} is Actively on intervention or is blank
\item \code{completed}: Participants whose \code{intervention_status} is Successfully completed
\item \code{withdrawn_testing}: Participants whose \code{intervention_status} is Withdrawn willing to complete testing
\item \code{withdrawn_no_testing}: Participants whose \code{intervention_status} is Withdrawn unwilling to complete testing
\item \code{lost}: Participants whose \code{intervention_status} is Lost to Follow-up
\item \code{group_}: A dataframe is returned for participants in each group. They are titled using the value of the group variable, i.e. group_(value)
}
}
\description{
This function transforms data from \code{\link[=get_data]{get_data()}} into consort ready
data by sorting participants into groups, i.e. screening, screen-failed, enrolled, etc.
}
\section{Notes}{

20240415: Might want to change the screen fail string to add line breaks every 30 characters.

All screened participants are filtered into statuses
}

