% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{screening_consort}
\alias{screening_consort}
\title{Create Screening Consort Image}
\usage{
screening_consort(consort_data = NULL, filepath = NULL)
}
\arguments{
\item{consort_data}{A list of dataframes from \code{\link[=prep_consort]{prep_consort()}}
\itemize{
\item \code{screened}: All screened participants that have a \code{scrn_date}
\item \code{not_screened}: Participants in the database without a \code{scrn_date}, presumably to be screened
\item \code{screen_failed}: Participants who screen-fail before consenting, \code{screening_status} = "Screen-Fail"
\item \code{eligible}: Participants who pass pre-consent screening, \code{screening_status} = "Eligible"
\item \code{on_hold}: Participants who are on-hold before consent, \code{screening_status} = On-hold
\item \code{still_screening}: Participants who are presumable still screening because \code{scrn_date} is not blank but \code{screening_status} is blank
\item \code{screen_fail_string}: String of pre-consent screenfail reasons and numbers for the consort diagram
}}

\item{filepath}{A file path to save the screening consort diagram image

@returns plot: a ggplot object

@import dplyr
@import ggplot2}
}
\description{
This function uses the dataframes created in \code{\link[=prep_consort]{prep_consort()}} to create a
screening consort diagram. THe screening consort image contains number screened, screen-failed
still screening, and eligible.
}
